% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bcoef.R
\name{Bcoef}
\alias{Bcoef}
\title{Coefficient matrix of an estimated VAR(p)}
\usage{
Bcoef(x)
}
\arguments{
\item{x}{An object of class \sQuote{\code{varest}}, generated by
\command{VAR()}.}
}
\value{
A matrix holding the estimated coefficients of a VAR.
}
\description{
Returns the estimated coefficients of a VAR(p) as a matrix.
}
\details{
Given an estimated VAR of the form: \deqn{ \hat{\bold{y}}_t = \hat{A}_1
\bold{y}_{t-1} + \ldots + \hat{A}_p \bold{y}_{t-p} + \hat{C}D_t } the
function returns the matrices \eqn{(\hat{A}_1 | \ldots | \hat{A}_p |
\hat{C})} as a matrix object.
}
\examples{

data(Canada)
var.2c <- VAR(Canada, p = 2, type = "const")
Bcoef(var.2c)

}
\seealso{
\code{\link{Acoef}}, \code{\link{VAR}}
}
\author{
Bernhard Pfaff
}
\keyword{regression}
