% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causality.R
\name{causality}
\alias{causality}
\title{Causality Analysis}
\usage{
causality(x, cause = NULL, vcov. = NULL, boot = FALSE, boot.runs = 100)
}
\description{
Computes the test statistics for Granger- and Instantaneous causality for a
VAR(p).
}
