% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.vec2var.R
\name{fitted.vec2var}
\alias{fitted.vec2var}
\alias{fitted}
\alias{fitted.values}
\alias{fitted.varest}
\title{Fit method for objects of class varest or vec2var}
\usage{
\method{fitted}{vec2var}(object, ...)
}
\arguments{
\item{object}{An object of class \sQuote{\code{varest}}; generated by
\command{VAR()}, or an object of class \sQuote{\code{vec2var}}; generated by
\command{vec2var()}.}

\item{\dots}{Currently not used.}
}
\description{
Returns the fitted values of a VAR(p)-model for objects generated by
\command{VAR()} or \command{vec2var()}. For objects of class varest the
fitted.values-method is applied to the list element \code{varresult}, which
is itself a list of \code{lm}-objects.
}
\examples{

data(Canada)
var.2c <- VAR(Canada, p = 2, type = "const")
fitted(var.2c)

}
\references{
Hamilton, J. (1994), \emph{Time Series Analysis}, Princeton
University Press, Princeton.

Lütkepohl, H. (2006), \emph{New Introduction to Multiple Time Series
Analysis}, Springer, New York.
}
\seealso{
\code{\link{VAR}}, \code{\link{vec2var}}
}
\author{
Bernhard Pfaff
}
\keyword{regression}
