% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_plot.lpirfs_lin_iv_obj.R
\name{ggplot.lpirfs_lin_iv_obj}
\alias{ggplot.lpirfs_lin_iv_obj}
\title{Plot impulse responses of lp_lin_iv in lpirfs}
\usage{
\method{ggplot}{lpirfs_lin_iv_obj}(
  irf,
  main = NULL,
  sub = NULL,
  cap = NULL,
  var_name = NULL,
  dev_new = FALSE,
  ...
)
}
\description{
Plot impulse responses of lp_lin_iv in lpirfs
}
\references{
Victor Espinoza, (2022), 
"Plot() impulse response function - show more than one in one window?,"
stackoverflow. \url{https://stackoverflow.com/questions/68010256/plot-impulse-response-function-show-more-than-one-in-one-window}
}
\author{
Victor Espinoza (stackoverflow), Koichi (Koiti) Yano
}
