% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.lpirfs_lin_obj.R
\name{ggplot.lpirfs_lin_obj}
\alias{ggplot.lpirfs_lin_obj}
\title{Plot impulse responses of lp_lin in lpirfs}
\usage{
\method{ggplot}{lpirfs_lin_obj}(
  irf,
  main = NULL,
  sub = NULL,
  cap = NULL,
  var_name = NULL,
  dev_new = FALSE,
  ...
)
}
\arguments{
\item{irf}{impulse responses of lp_lin in lpirfs}

\item{main}{main title of plot (The default is NULL, in which case the main
title is generated automatically.)}

\item{sub}{subtitile of plot}

\item{cap}{caption of plot}

\item{var_name}{variable names: ex. var_name=c("Emp", "Prod", "RW", "Unemp")}

\item{dev_new}{logical. If TRUE, open a new graphics device.}

\item{\dots}{further arguments passed to or from other methods 
(currently not used).}
}
\value{
A ggplot object
}
\description{
Plot impulse responses of lp_lin in lpirfs
}
\examples{
\dontrun{
require(lpirfs)
require(vars2)
# Load (endogenous) data
endog_data <- interest_rules_var_data
# Estimate linear model
lp_irf <- lp_lin(endog_data, lags_endog_lin = 4, trend = 0,
                 shock_type = 1, confint  = 1.96, hor = 12)
# Compare with Figure 5 in Jordà (2005)
ggplot(lp_irf)
vars_res <- VAR(endog_data, p=2)
vars_irf <- irf(vars_res)
ggplot(vars_irf)  
}
}
\references{
Victor Espinoza, (2022), 
"Plot() impulse response function - show more than one in one window?,"
stackoverflow. \url{https://stackoverflow.com/questions/68010256/plot-impulse-response-function-show-more-than-one-in-one-window}
}
\author{
Victor Espinoza (stackoverflow), Koichi (Koiti) Yano
}
