% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phi.varest.R
\name{Phi.varest}
\alias{Phi.varest}
\title{Phi.varest:}
\usage{
\method{Phi}{varest}(x, nstep = 10, ...)
}
\arguments{
\item{x}{An object of class \sQuote{\code{varest}}, generated by
\command{VAR()}, or an object of class \sQuote{\code{svarest}}, generated by
\command{SVAR()}, or an object of class \sQuote{\code{svecest}}, generated
by \command{SVEC()}, or an object of class \sQuote{\code{vec2var}},
generated by \command{vec2var()}.}

\item{nstep}{An integer specifying the number of moving error coefficient
matrices to be calculated.}

\item{...}{Currently not used.}
}
\description{
Phi.varest:
}
