% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.varirf.R
\name{plot.varirf}
\alias{plot.varirf}
\title{residuals.varest}
\usage{
\method{plot}{varirf}(
  x,
  plot.type = c("multiple", "single"),
  names = NULL,
  main = NULL,
  sub = NULL,
  lty = NULL,
  lwd = NULL,
  col = NULL,
  ylim = NULL,
  ylab = NULL,
  xlab = NULL,
  nc,
  mar.multi = c(0, 4, 0, 4),
  oma.multi = c(6, 4, 6, 4),
  adj.mtext = NA,
  padj.mtext = NA,
  col.mtext = NA,
  ...
)
}
\description{
residuals.varest
}
