% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.svarest.R
\name{print.svarest}
\alias{print.svarest}
\title{Print methods for objects in SVAR}
\usage{
\method{print}{svarest}(x, digits = max(3, getOption("digits") - 3), ...)
}
\description{
Print methods for objects in SVAR
}
