% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.varest.R
\name{residuals.varest}
\alias{residuals.varest}
\title{residuals.varest}
\usage{
\method{residuals}{varest}(object, ...)
}
\arguments{
\item{object}{An object of class ‘varest’; generated by VAR(), 
or an object of class ‘vec2var’; generated by vec2var()}

\item{\dots}{Currently not used.}
}
\description{
residuals.varest
}
