% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_plot.lpirfs_lin_obj.R
\name{vars_plot.lpirfs_lin_obj}
\alias{vars_plot.lpirfs_lin_obj}
\title{Plot impulse responses of lp_lin in lpirfs}
\usage{
\method{vars_plot}{lpirfs_lin_obj}(
  irf,
  main = NULL,
  sub = NULL,
  cap = NULL,
  var_name = NULL,
  dev_new = FALSE,
  ...
)
}
\arguments{
\item{irf}{impulse responses of lp_lin in lpirfs}

\item{main}{main title of plot (The default is NULL, in which case the main
title is generated automatically.)}

\item{sub}{subtitile of plot}

\item{cap}{caption of plot}

\item{var_name}{variable names: ex. var_name=c("Emp", "Prod", "RW", "Unemp")}

\item{dev_new}{logical. If TRUE, open a new graphics device.}

\item{\dots}{further arguments passed to or from other methods 
(currently not used).}
}
\value{
A ggplot object
}
\description{
Plot impulse responses of lp_lin in lpirfs
}
\examples{
\dontrun{
require(vars2)
require(lpirfs)

lp_p2_irf <- lp_lin(endog_data= Canada_tbl[,2:5], lags_endog_lin = 2,
trend = 0, shock_type  = 1, confint = 1.96, hor = 10)
# Show all impule responses
vars_plot(lp_p2_irf, dev_new=TRUE)
}
}
\author{
Koichi (Koiti) Yano
}
