% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_plot.R
\name{vars_plot}
\alias{vars_plot}
\title{Plot impulse responses with ggplot2}
\usage{
vars_plot(irf, ...)
}
\arguments{
\item{irf}{impulse responses of VAR or SVAR}

\item{\dots}{further arguments passed to or from other methods 
(currently not used).}
}
\description{
Plot impulse responses with ggplot2
}
