% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARselect.R
\name{VARselect}
\alias{VARselect}
\title{VARselect: Information criteria for different VAR(p)}
\usage{
VARselect(
  y,
  lag.max = 10,
  type = c("const", "trend", "both", "none"),
  season = NULL,
  exogen = NULL
)
}
\arguments{
\item{y}{Data item containing the endogenous variables}

\item{lag.max}{Integer, determines the highest lag order for lag length selection according to the choosen ic}

\item{type}{Type of deterministic regressors to include.}

\item{season}{Inlusion of centered seasonal dummy variables (integer value of frequency).}

\item{exogen}{Inlusion of exogenous variables.}
}
\value{
list
}
\description{
VARselect: Information criteria for different VAR(p)
}
