% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggplot}
\alias{ggplot}
\title{Plot method with ggplot2}
\usage{
ggplot(irf, ...)
}
\arguments{
\item{irf}{impulse responses of VAR or SVAR}

\item{\dots}{further arguments passed to or from other methods 
(currently not used).}
}
\description{
Plot method with ggplot2
}
