% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.varirf.R
\name{ggplot.varirf}
\alias{ggplot.varirf}
\title{ggplot.varirf: Plot impulse responses of VAR and SVAR with ggplot2}
\usage{
\method{ggplot}{varirf}(irf)
}
\arguments{
\item{irf}{a list of impulse responses}
}
\value{
plot
}
\description{
ggplot.varirf: Plot impulse responses of VAR and SVAR with ggplot2
}
\references{
Victor Espinoza, (2022),
"Plot() impulse response function - show more than one in one window?,"
stackoverflow.
}
\author{
Victor Espinoza (stackoverflow), Koichi (Koiti) Yano
}
