% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.varirf.R
\name{ggplot.varirf}
\alias{ggplot.varirf}
\title{Plot impulse responses of VAR and SVAR}
\usage{
\method{ggplot}{varirf}(irf, main = NULL, sub = NULL, cap = NULL, ...)
}
\arguments{
\item{irf}{impulse responses of VAR or SVAR}

\item{main}{main title of plot (The default is NULL, in which case the main
title is generated automatically.)}

\item{sub}{subtitile of plot}

\item{cap}{caption of plot}

\item{\dots}{further arguments passed to or from other methods 
(currently not used).}
}
\description{
Plot impulse responses of VAR and SVAR
}
\examples{
data(Canada)
## For VAR
var.2c <- VAR(Canada, p = 2, type = "const")
irf.2c <- irf(var.2c, impulse = "e", response = c("e", "prod", "rw", "U"), boot =
TRUE)
ggplot(irf.2c, sub="Canada", cap="Caption")

## For SVAR
amat <- diag(4)
diag(amat) <- NA
svar.a <- SVAR(var.2c, estmethod = "direct", Amat = amat)
irf.sa <- irf(svar.a)
ggplot(irf.sa, main="Canada", sub="Structural IRF", 
cap="Caption: The original time series are published by the OECD.")
}
\references{
Victor Espinoza, (2022), 
"Plot() impulse response function - show more than one in one window?,"
stackoverflow.
}
\author{
Victor Espinoza (on stackoverflow), Koichi (Koiti) Yano
}
