% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union.R
\name{union}
\alias{union}
\title{Unite data from API (ex. FRED)}
\usage{
union(data_list, var_names, ...)
}
\arguments{
\item{data_list}{impulse responses of VAR or SVAR}

\item{var_names}{subtitile of plot (main title is generated automatically)}

\item{\dots}{further arguments passed to or from other methods 
(currently not used).}
}
\description{
Unite data from API (ex. FRED)
}
\examples{
\dontrun{
require("fredr")
library(tidyverse)
# FREDのAPIを利用するキーを指定します
fredr_set_key("d88c682c8274047b72e7e64da46a7801")
#freq_flag <- "m" # Monthly
freq_flag <- "q" # Quarterly
#freq_flag <- "a" # Annual
# In FRED, the start and the end of the observation period
start1_date <- "2018-01-01"
end1_date <- Sys.Date()

# https://fred.stlouisfed.org/series/NAEXKP02JPQ189S
jp_prv_c <- fredr_series_observations(
series_id = "NAEXKP02JPQ189S",
observation_start = as.Date(start1_date),
observation_end = as.Date(end1_date),
frequency = freq_flag,
units = "log"
)

# https://fred.stlouisfed.org/series/NAEXKP03JPQ189S
jp_gov_e <- fredr_series_observations(
series_id = "NAEXKP03JPQ189S",
observation_start = as.Date(start1_date),
observation_end = as.Date(end1_date),
frequency = freq_flag,
units = "log"
)

# https://fred.stlouisfed.org/series/JPNGDPDEFQISMEI
jp_def <- fredr_series_observations(
series_id = "JPNGDPDEFQISMEI",
observation_start = as.Date(start1_date),
observation_end = as.Date(end1_date),
frequency = freq_flag,
units = "log"
)

# https://stackoverflow.com/questions/36520813/r-dplyr-rename-and-select-using-string-variable
fredr_list <- list(jp_prv_c, jp_gov_e, jp_def)
ar_names <- c("priv_c", "gov_e", "defl")
union.fredr(fredr_list, var_names) -> jp_macro
print(jp_macro)
}
}
