% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_plot.varirf.R
\name{vars_plot.varirf}
\alias{vars_plot.varirf}
\title{Plot impulse responses of VAR and SVAR}
\usage{
\method{vars_plot}{varirf}(
  irf,
  main = NULL,
  sub = NULL,
  cap = NULL,
  resp_name = NULL,
  imp_name = NULL,
  dev_new = FALSE,
  graph_style = "pw",
  ...
)
}
\arguments{
\item{irf}{impulse responses of VAR or SVAR}

\item{main}{main title of plot (not implemented yet)}

\item{sub}{subtitile of plot (not implemented yet)}

\item{cap}{caption of plot (not implemented yet)}

\item{resp_name}{variable names of reponse: ex. resp_name=c("Emp", "Prod", "RW", "Unemp")}

\item{imp_name}{variable names of impulse: ex. imp_name=c("Emp")}

\item{dev_new}{logical. If TRUE, open a new graphics device.}

\item{\dots}{further arguments passed to or from other methods 
(currently not used).}
}
\value{
A ggplot object
}
\description{
Plot impulse responses of VAR and SVAR
}
\examples{
\donttest{
data(Canada)
## For VAR
var.2c <- VAR(Canada, p = 2, type = "const")
irf.2c <- irf(var.2c, impulse = "e", response = c("e", "prod", "rw", "U"), boot =
TRUE)
vars_plot(irf.2c, sub="Canada", cap="Caption", imp_name=c("Emp"),
resp_name=c("Emp", "Prod", "Real Wage", "Unemp"))

## For SVAR
amat <- diag(4)
diag(amat) <- NA
svar.a <- SVAR(var.2c, estmethod = "direct", Amat = amat)
irf.sa <- irf(svar.a)
vars_plot(irf.sa, main="Canada", sub="Structural IRF", 
cap="Caption: The original time series are published by the OECD.",
var_name=c("Emp", "Prod", "Real Wage", "Unemp"), dev_new=TRUE)
}
}
\references{
Victor Espinoza, (2022), 
"Plot() impulse response function - show more than one in one window?,"
stackoverflow. \url{https://stackoverflow.com/questions/68010256/plot-impulse-response-function-show-more-than-one-in-one-window}
}
\author{
Victor Espinoza (stackoverflow), Koichi (Koiti) Yano
}
